pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "alltius-pdf-extractor"
version = "0.1.0"
description = "PDF to structured JSON extractor (paragraphs, tables, charts)"
readme = "README.md"
authors = [{name = "Alltius", email = "dev@alltius.ai"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
  "PyMuPDF==1.24.9",
  "pdfplumber==0.11.4",
  "camelot-py[cv]==0.11.0",
  "pandas==2.2.2",
  "pillow>=10.0.0",
  "pytesseract>=0.3.10"
]

[project.optional-dependencies]
 dev = ["pytest", "reportlab", "pillow", "pytesseract"]

[project.scripts]
alltius-extract = "alltius_ai.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
